Improved: order detail page UI using css grid, removed the the usage of isPlatform of ionic instead used ion-hide class and removed the orderInfo component (#283)
