+++ src/locales/en.json
@@ -74,7 +74,8 @@
   "Order delivered to": "Order delivered to { customerName }",
   "Order details": "Order details",
   "Order edit permissions": "Order edit permissions",
-  "Order is now ready for handover.": "Order is now ready for handover.",
+  "Order is now ready to handover.": "Order is now ready to handover.",
+  "Order has been rejected.": "Order has been rejected.",
   "Order marked as ready for pickup, an email notification has been sent to the customer": "Order marked as ready for pickup, an email notification has been sent to the customer",
   "Order marked as ready for pickup but something went wrong while sending the email notification": "Order marked as ready for pickup but something went wrong while sending the email notification",
   "Order not found": "Order not found",
