+++ src/locales/es.json
@@ -73,7 +73,8 @@
   "Order delivered to": "Orden entregada a {customerName}",
   "Order details": "Detalles de la orden",
   "Order edit permissions": "Permisos de edición de la orden",
-  "Order is now ready for handover.": "Order is now ready for handover.",
+  "Order is now ready to handover.": "Order is now ready to handover.",
+  "Order has been rejected.": "Order has been rejected.",
   "Order marked as ready for pickup, an email notification has been sent to the customer": "Pedido marcado como listo para recoger, se ha enviado una notificación por correo electrónico al cliente",
   "Order marked as ready for pickup but something went wrong while sending the email notification": "Pedido marcado como listo para recoger pero algo salió mal al enviar la notificación por correo electrónico",
   "Order not found": "Orden no encontrada",
