+++ src/locales/ja.json
@@ -73,7 +73,8 @@
   "Order delivered to": "注文は { customerName }様 に配達されました",
   "Order details": "注文詳細",
   "Order edit permissions": "注文の編集権限",
-  "Order is now ready for handover.": "Order is now ready for handover.",
+  "Order is now ready to handover.": "Order is now ready to handover.",
+  "Order has been rejected.": "Order has been rejected.",
   "Order marked as ready for pickup, an email notification has been sent to the customer": "注文された商品の受取準備完了とマークされ、Eメールによる通知がお客様に送信されました。",
   "Order marked as ready for pickup but something went wrong while sending the email notification": "注文された商品の受取準備完了とマークされましたが、Eメールによる通知の送信中に何かしらの問題が発生しました。",
   "Order not found": "注文が見つかりません",
