+++ src/views/OrderDetail.vue
@@ -5,35 +5,66 @@
         <ion-back-button default-href="/" slot="start" />
         <ion-title>{{ $t("Order details") }}</ion-title>
         <ion-buttons slot="end">
-          <ion-button @click="openOrderItemRejHistoryModal()">
+          <ion-button :disabled="order?.rejected" @click="openOrderItemRejHistoryModal()">
             <ion-icon slot="icon-only" :icon="timeOutline" />
           </ion-button>
-          <ion-button v-if="!isDesktop" :disabled="!hasPermission(Actions.APP_ORDER_UPDATE) || order?.readyToHandover || order?.rejected" @click="updateOrder(order)">
+          <ion-button class="ion-hide-md-up" :disabled="!hasPermission(Actions.APP_ORDER_UPDATE) || order?.readyToHandover || order?.rejected" @click="updateOrder(order)">
             <ion-icon slot="icon-only" color="danger" :icon="bagRemoveOutline" />
           </ion-button>
         </ion-buttons>
       </ion-toolbar>
     </ion-header>
+
     <ion-content>
-      <main :class="{ 'desktop-only' : isDesktop && order?.part?.items?.length }">
-        <OrderInfo v-if="!isDesktop" />
+      <main>
+        <aside>
+          <ion-item v-if="order?.readyToHandover || order?.rejected" color="light" lines="none">
+            <ion-icon :icon="order?.readyToHandover ? checkmarkCircleOutline : order?.rejected ? closeCircleOutline : ''" :color="order?.readyToHandover ? 'success' : order?.rejected ? 'danger' : ''" slot="start" />
+            <ion-label class="ion-text-wrap">{{ order?.readyToHandover ? $t("Order is now ready to handover.") : order?.rejected ? $t("Order has been rejected.") : '' }}</ion-label>
+          </ion-item>
+          <ion-list>
+            <ion-item lines="none">
+              <ion-label class="ion-text-wrap">
+                <h2>{{ order?.customer?.name }}</h2>
+                <p>{{ order?.orderName ? order?.orderName : order?.orderId }}</p>
+              </ion-label>
+              <ion-badge v-if="order.placedDate" slot="end">{{ timeFromNow(order.placedDate) }}</ion-badge>
+            </ion-item>
+          </ion-list>
+          <ion-item v-if="customerEmail" lines="none">
+            <ion-icon :icon="mailOutline" slot="start" />
+            <ion-label>{{ customerEmail }}</ion-label>
+            <ion-icon :icon="copyOutline" slot="end" @click="copyToClipboard(customerEmail)" />
+          </ion-item>
+          <ion-item v-if="order?.shippingInstructions" color="light" lines="none">
+            <ion-label class="ion-text-wrap">
+              <p class="overline">{{ $t("Handling Instructions") }}</p>
+              <p>{{ order?.shippingInstructions }}</p>
+            </ion-label>
+          </ion-item>
+          <div class="ion-margin-top ion-hide-md-down">
+            <!-- TODO: implement functionality to change shipping address -->
+            <ion-button :disabled="!hasPermission(Actions.APP_ORDER_UPDATE) || order?.readyToHandover || order?.rejected" expand="block" @click.stop="readyForPickup(order, order.part)">
+              {{ order?.part?.shipmentMethodEnum?.shipmentMethodEnumId === 'STOREPICKUP' ? $t("Ready for pickup") : $t("Ready to ship") }}
+            </ion-button>
+            <ion-button :disabled="!hasPermission(Actions.APP_ORDER_UPDATE) || order?.readyToHandover || order?.rejected" expand="block" color="danger" fill="outline" @click="updateOrder(order)">
+              {{ $t("Reject Order") }}
+            </ion-button>
+          </div>
+        </aside>
         <section>
           <ion-card v-for="(item, index) in order.part?.items" :key="index">
             <ProductListItem :item="item" />
-            <ion-item v-if="partialOrderRejection" lines="none" class="border-top">
+            <div v-if="partialOrderRejection" class="border-top">
               <ion-button fill="clear" @click="openReportAnIssueModal(item)">
                 {{ $t("Report an issue") }}
               </ion-button>
-            </ion-item>
+            </div>
           </ion-card>
         </section>
-
-        <aside v-if="isDesktop">
-          <OrderInfo />
-        </aside>
       </main>
 
-      <ion-fab v-if="!isDesktop" vertical="bottom" horizontal="end" slot="fixed" @click="readyForPickup({ order, part: order.part })">
+      <ion-fab class="ion-hide-md-up" vertical="bottom" horizontal="end" slot="fixed" @click="readyForPickup(order, order.part)">
         <ion-fab-button :disabled="!hasPermission(Actions.APP_ORDER_UPDATE) || order?.readyToHandover || order?.rejected">
           <ion-icon :icon="bagHandleOutline" />
         </ion-fab-button>
@@ -46,6 +77,7 @@
 import {
   alertController,
   IonBackButton,
+  IonBadge,
   IonButton,
   IonButtons,
   IonCard,
@@ -53,30 +85,44 @@ import {
   IonHeader,
   IonIcon,
   IonItem,
+  IonList,
   IonPage,
+  IonLabel,
   IonTitle,
   IonToolbar,
   IonFab,
   IonFabButton,
   modalController,
-  isPlatform,
 } from "@ionic/vue";
 import { defineComponent } from "vue";
 import { mapGetters, useStore } from "vuex";
-import { bagHandleOutline, bagRemoveOutline, timeOutline } from "ionicons/icons";
+import {
+  copyOutline,
+  closeCircleOutline,
+  checkmarkCircleOutline,
+  mailOutline,
+  sendOutline,
+  bagHandleOutline,
+  bagRemoveOutline,
+  timeOutline
+} from "ionicons/icons";
 import ProductListItem from '@/components/ProductListItem.vue'
 import { useRouter } from 'vue-router'
 import { Actions, hasPermission } from '@/authorization'
 import OrderItemRejHistoryModal from '@/components/OrderItemRejHistoryModal.vue';
 import ReportAnIssueModal from '@/components/ReportAnIssueModal.vue';
-import OrderInfo from '@/components/OrderInfo.vue';
-import emitter from "@/event-bus"
 import AssignPickerModal from "@/views/AssignPickerModal.vue";
+import { copyToClipboard } from '@/utils'
+import { DateTime } from "luxon";
+import { hasError } from '@/adapter';
+import ShipToCustomerModal from "@/components/ShipToCustomerModal.vue";
+import { OrderService } from "@/services/OrderService";
 
 export default defineComponent({
   name: "OrderDetail",
   components: {
     IonBackButton,
+    IonBadge,
     IonButton,
     IonButtons,
     IonCard,
@@ -84,17 +130,18 @@ export default defineComponent({
     IonHeader,
     IonIcon,
     IonItem,
+    IonList,
     IonPage,
+    IonLabel,
     IonTitle,
     IonToolbar,
     ProductListItem,
-    OrderInfo,
     IonFab,
     IonFabButton,
   },
-  data () {
+  data() {
     return {
-      isDesktop: isPlatform('desktop')
+      customerEmail: ''
     }
   },
   computed: {
@@ -106,12 +153,25 @@ export default defineComponent({
     })
   },
   methods: {
-    async assignPicker(order: any, part: any, facilityId: any) {      
+    async assignPicker(order: any, part: any, facilityId: any) {
       const assignPickerModal = await modalController.create({
         component: AssignPickerModal,
         componentProps: { order, part, facilityId }
       });
       return assignPickerModal.present();
+    },
+    async openOrderItemRejHistoryModal() {
+      const orderItemRejHistoryModal = await modalController.create({
+        component: OrderItemRejHistoryModal,
+      });
+      return orderItemRejHistoryModal.present();
+    },
+    async openReportAnIssueModal(item: any) {
+      const reportAnIssueModal = await modalController.create({
+        component: ReportAnIssueModal,
+        componentProps: { item }
+      });
+      return reportAnIssueModal.present();
     },
     async getOrderDetail(orderId: any, orderPartSeqId: any) {
       const payload = {
@@ -135,7 +195,7 @@ export default defineComponent({
               this.store.dispatch('order/setUnfillableOrderOrItem', { orderId: order.orderId, part: order.part }).then((resp) => {
                 if (resp) {
                   // Mark current order as rejected
-                  this.store.dispatch('order/updateCurrent', { order : { ...order, rejected: true } })
+                  this.store.dispatch('order/updateCurrent', { order: { ...order, rejected: true } })
                 }
               })
             },
@@ -143,24 +203,11 @@ export default defineComponent({
         });
       return alert.present();
     },
-    async openOrderItemRejHistoryModal(){
-      const orderItemRejHistoryModal = await modalController.create({
-        component: OrderItemRejHistoryModal,
-      });
-      return orderItemRejHistoryModal.present();
-    },
-    async openReportAnIssueModal(item: any){
-      const reportAnIssueModal = await modalController.create({
-        component: ReportAnIssueModal,
-        componentProps: { item }
-      });
-      return reportAnIssueModal.present();
-    },
-    async readyForPickup (orderInfo: any) {
-      if(this.configurePicker) return this.assignPicker(orderInfo.order, orderInfo.part, this.currentFacility.facilityId);
-      const pickup = orderInfo.part?.shipmentMethodEnum?.shipmentMethodEnumId === 'STOREPICKUP';
+    async readyForPickup(order: any, part: any) {
+      if (this.configurePicker) return this.assignPicker(order, part, this.currentFacility.facilityId);
+      const pickup = part?.shipmentMethodEnum?.shipmentMethodEnumId === 'STOREPICKUP';
       const header = pickup ? this.$t('Ready for pickup') : this.$t('Ready to ship');
-      const message = pickup ? this.$t('An email notification will be sent to that their order is ready for pickup. This order will also be moved to the packed orders tab.', { customerName: orderInfo.order.customer.name, space: '<br/><br/>'}) : '';
+      const message = pickup ? this.$t('An email notification will be sent to that their order is ready for pickup. This order will also be moved to the packed orders tab.', { customerName: order.customer.name, space: '<br/><br/>' }) : '';
 
       const alert = await alertController
         .create({
@@ -169,10 +216,10 @@ export default defineComponent({
           buttons: [{
             text: this.$t('Cancel'),
             role: 'cancel'
-          },{
+          }, {
             text: header,
             handler: async () => {
-              await this.store.dispatch('order/packShipGroupItems', { order: orderInfo.order, part: orderInfo.part, facilityId: this.currentFacility.facilityId})
+              await this.store.dispatch('order/packShipGroupItems', { order: order, part: part, facilityId: this.currentFacility.facilityId })
             }
           }]
         });
@@ -180,19 +227,34 @@ export default defineComponent({
     },
     async fetchRejectReasons() {
       await this.store.dispatch('util/fetchRejectReasons');
+    },
+    timeFromNow(time: string) {
+      const timeDiff = DateTime.fromISO(time).diff(DateTime.local());
+      return DateTime.local().plus(timeDiff).toRelative();
+    },
+    async shipToCustomer() {
+      const shipmodal = await modalController.create({
+        component: ShipToCustomerModal,
+      });
+      return shipmodal.present();
+    },
+    async getCustomerContactDetails() {
+      try {
+        const resp = await OrderService.getCustomerContactDetails(this.$route.params.orderId)
+        if (!hasError(resp)) {
+          this.customerEmail = resp.data.orderContacts.email.email
+        }
+      } catch (error) {
+        console.error(error)
+      }
     }
-   },
+  },
   async mounted() {
     await this.getOrderDetail(this.$route.params.orderId, this.$route.params.orderPartSeqId);
+    await this.getCustomerContactDetails()
     await this.fetchRejectReasons();
-    emitter.on('updateOrder', this.updateOrder);
-    emitter.on('readyForPickupOfOrderDetail', this.readyForPickup);
   },
-  unmounted() {
-    emitter.off('updateOrder', this.updateOrder);
-    emitter.off('readyForPickupOfOrderDetail', this.readyForPickup);
-  },
-  setup () {
+  setup() {
     const store = useStore();
     const router = useRouter();
 
@@ -200,10 +262,16 @@ export default defineComponent({
       Actions,
       bagHandleOutline,
       bagRemoveOutline,
+      copyOutline,
+      copyToClipboard,
+      closeCircleOutline,
+      checkmarkCircleOutline,
       hasPermission,
       router,
       store,
-      timeOutline
+      timeOutline,
+      mailOutline,
+      sendOutline
     };
   }
 });
@@ -214,27 +282,25 @@ export default defineComponent({
   border-top: 1px solid #ccc;
 }
 
-/* If not in desktop */
-main {
-  max-width: 445px;
-  margin: var(--spacer-base) auto 0; 
-}
+@media (min-width: 768px) {
+  main {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    grid-template-areas: 'left right';
+    gap: var(--spacer-base);
+    margin: var(--spacer-lg) var(--spacer-lg) 0;
+  }
 
-aside {
-  position: sticky;
-  top: var(--spacer-lg);
-}
+  aside {
+    grid-area: right;
+    /* Order Info section to stay fixed and not scrollable as there can be multiple order items which will increase the grid container size */
+    height: fit-content; 
+    position: sticky;
+    top: var(--spacer-lg);
+  }
 
-.desktop-only {
-  display: flex;
-  justify-content: center;
-  align-items: start;
-  gap: var(--spacer-2xl);
-  max-width: 990px;
-  margin: var(--spacer-base) auto 0;
-}
-
-.desktop-only > * {
-  width: calc(50% - var(--spacer-xl));
+  section {
+    grid-area: left;
+  }
 }
 </style>
